move from ./pkg directory to the ./internal directory
