+++ internal/handler/handler.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/Dyleme/image-coverter/pkg/model"
+	"github.com/Dyleme/image-coverter/internal/model"
 	"github.com/gorilla/mux"
 	"github.com/sirupsen/logrus"
 )
