+++ internal/handler/requests.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/Dyleme/image-coverter/pkg/jwt"
-	"github.com/Dyleme/image-coverter/pkg/model"
+	"github.com/Dyleme/image-coverter/internal/jwt"
+	"github.com/Dyleme/image-coverter/internal/model"
 	"github.com/gorilla/mux"
 )
