+++ internal/handler/response_functions.go
