+++ internal/repository/download_postgres_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
-	"github.com/Dyleme/image-coverter/pkg/repository"
+	"github.com/Dyleme/image-coverter/internal/repository"
 )
 
 func TestGetImageUrl(t *testing.T) {
