+++ internal/service/auth_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Dyleme/image-coverter/pkg/model"
-	"github.com/Dyleme/image-coverter/pkg/service"
-	"github.com/Dyleme/image-coverter/pkg/service/mocks"
+	"github.com/Dyleme/image-coverter/internal/model"
+	"github.com/Dyleme/image-coverter/internal/service"
+	"github.com/Dyleme/image-coverter/internal/service/mocks"
 	"github.com/golang/mock/gomock"
 )
