+++ internal/service/mocks/mock_auth.go
@@ -8,7 +8,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	model "github.com/Dyleme/image-coverter/pkg/model"
+	model "github.com/Dyleme/image-coverter/internal/model"
 	gomock "github.com/golang/mock/gomock"
 )
